fixed addproduct page, shows products on products page and shows your own products on profile page
