+++ src/components/ProductComponent/ProductCard.tsx
@@ -6,63 +6,79 @@ import {
   CardContent,
   Typography,
 } from "@mui/material";
-import CarImage from "../../assets/Rectangle42.png";
-import PlaceIcon from "@mui/icons-material/Place";
+import {onAuthStateChanged} from "firebase/auth";
+import {useState, useEffect} from "react";
+import {useAuth} from "../../backend/Context";
+import {auth} from "../../backend/firebase";
+
 function ProductCard() {
+  const {getProductsByUserID, createNotification} = useAuth();
+
+  const [products, setProducts] = useState<any[]>([]);
+
+  useEffect(() => {
+    onAuthStateChanged(auth, (user) => {
+      if (user) {
+        getProductsByUserID(user.uid).then((products) => {
+          setProducts(products);
+        });
+      }
+    });
+  }, [getProductsByUserID]);
   return (
     <>
-      <Box
-        sx={{
-          display: "flex",
-          "@media screen and (max-width: 600px)": {display: "none"},
-        }}
-      >
-        <Card sx={{width: "227px", height: "330px", m: "0.3rem"}}>
-          <img style={{width: "100%", height: "173px"}} src={CarImage} alt="" />
-          <CardContent>
-            <Typography
-              sx={{fontFamily: "sans-serif"}}
-              aria-label="Medium sizes"
-              gutterBottom
-              component="div"
-            >
-              La Ferrari Pasta racer 2022
-            </Typography>
-          </CardContent>
-          <Typography variant="body2" color="text.secondary">
-            <PlaceIcon sx={{color: "pink"}} /> västra göteborg
-          </Typography>
-          <CardActions sx={{display: "flex", justifyContent: "space-between"}}>
-            <Typography sx={{ml: ".7rem"}}>230kr</Typography>
-            <Button
-              size="small"
-              sx={{
-                height: "1.5rem",
-                backgroundColor: "pink",
-                border: "1px solid white",
-                color: "white",
-                mr: ".7rem",
-                boxShadow: 3,
-              }}
-            >
-              Contact
-            </Button>
-          </CardActions>
-        </Card>
-      </Box>
-
-      <Box
-        sx={{
-          height: "200px",
-          width: "125px",
-          mr: ".5rem",
-          "@media screen and (min-width: 600px)": {display: "none"},
-        }}
-      >
-        <img style={{width: "125px"}} src={CarImage} alt="car" />
-        <Typography sx={{fontSize: "1rem"}}> La Ferrari Pasta </Typography>
-        <Typography variant="body2"> 230kr</Typography>
-      </Box>
+      {products.map((item) => {
+        return (
+          <Box
+            sx={{
+              display: "flex",
+              "@media screen and (max-width: 600px)": {display: "none"},
+            }}
+            key={item.id}
+          >
+            <Card sx={{width: "227px", height: "330px", m: "0.3rem"}}>
+              <img
+                style={{width: "100%", height: "173px"}}
+                src={item.image}
+                alt={item.title}
+              />
+              <CardContent>
+                <Typography
+                  sx={{fontFamily: "sans-serif"}}
+                  aria-label="Medium sizes"
+                  gutterBottom
+                  component="div"
+                >
+                  {item.title}
+                </Typography>
+              </CardContent>
+              <CardActions
+                sx={{display: "flex", justifyContent: "space-between"}}
+              >
+                <Typography sx={{ml: ".7rem"}}>{item.price} kr</Typography>
+                <Button
+                  size="small"
+                  sx={{
+                    height: "1.5rem",
+                    backgroundColor: "pink",
+                    border: "1px solid white",
+                    color: "white",
+                    mr: ".7rem",
+                    boxShadow: 3,
+                  }}
+                  onClick={() => {
+                    createNotification(item.id).then(() => {
+                      console.log("Create Notification done");
+                    });
+                  }}
+                >
+                  Contact
+                </Button>
+              </CardActions>
+            </Card>
+          </Box>
+        );
+      })}
     </>
   );
 }
