+++ src/components/ProductComponent/ProductPageComponent.tsx
@@ -24,12 +24,6 @@ function ProductPageComponent() {
       >
         <Box sx={{display: "flex"}}>
           <ProductCard />
-          <ProductCard />
-          <ProductCard />
-          <ProductCard />
-          <ProductCard />
-          <ProductCard />
-          <ProductCard />
         </Box>
       </ImageList>
